Extract codes to component DynamicSimulationResultChartTabs
