+++ src/components/results/dynamicsimulation/dynamic-simulation-result-chart.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -7,7 +7,7 @@
 
 import { Grid } from '@mui/material';
 import DynamicSimulationResultSeriesList from './dynamic-simulation-result-series-list';
-import { useMemo, useState } from 'react';
+import { memo, useMemo, useState } from 'react';
 import DynamicSimulationResultSeriesChart from './dynamic-simulation-result-series-chart';
 
 const headers = ['Left Axis', 'Available Curves', 'Right Axis'];
@@ -61,4 +61,4 @@ const DynamicSimulationResultChart = ({ series }) => {
     );
 };
 
-export default DynamicSimulationResultChart;
+export default memo(DynamicSimulationResultChart);
