+++ src/components/results/dynamicsimulation/dynamic-simulation-result-series-chart.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -23,7 +23,7 @@ const baseColors = [red, orange, blue, green, purple, pink, cyan];
 const PlotlySeriesChart = ({ leftSeries, rightSeries }) => {
     const data = useMemo(() => {
         function seriesToData(opts) {
-            return (s, index) => ({
+            return (s) => ({
                 name: s.name,
                 type: 'scatter',
                 mode: 'lines+markers',
